fix context Spring security context cleared out of creating external beans. The CustomUserDetailsService have their authentication manager set via setter. Persistent token repository injected.
Admin Controller: moved defaults to their @RequestParams definitions.

Smoke test now has assertion
