+++ src/test/java/pl/dreilt/basicspringmvcapp/BasicSpringMvcAppApplicationTests.java
@@ -2,12 +2,14 @@ package pl.dreilt.basicspringmvcapp;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.util.Assert;
 
 @SpringBootTest
 class BasicSpringMvcAppApplicationTests {
 
     @Test
     void contextLoads() {
+        Assert.isTrue(true, "Smoke test runs");
     }
 
 }
